+++ M1/Lat2/persegipanjang.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
-int p, l , luas, keliling;
+float p, l , luas, keliling;
 
 //Fungsi formula luas & keliling persegi panjang
-int lkPersegiPanjang(int p, int l, int* luas, int* keliling){
+float lkPersegiPanjang(float p, float l, float* luas, float* keliling){
     *luas = p * l;
     *keliling = 2 * (p + l);
 
@@ -13,9 +13,9 @@ int lkPersegiPanjang(int p, int l, int* luas, int* keliling){
 void apersegiPanjang(){
     printf("Inputkan nilai panjang dan lebar [dalam cm]\n");
     printf("Panjang : ");
-    scanf("%d", &p);
+    scanf("%f", &p);
     printf("\nLebar: ");
-    scanf("%d", &l);
+    scanf("%f", &l);
     lkPersegiPanjang(p, l, &luas, &keliling);
-    printf("\nLuas : %d dan Keliling : %d dengan panjang : %d  dan lebar : %d \n\n", luas, keliling, p, l);
+    printf("\nLuas : %.2f dan Keliling : %.2f dengan panjang : %.2f  dan lebar : %.2f \n\n", luas, keliling, p, l);
 }
